requirements.txt
baresipy
python-dotenv
PyQt5
pip
certifi
pexpect
ptyprocess
urllib3
opentone
pydub
idna
